Create BondingCurveGraph component
