+++ src/components/v1/FundingCycle/FundingCycleDetails.tsx
@@ -139,7 +139,16 @@ export default function FundingCycleDetails({
             label={
               <TooltipLabel
                 label={t`Bonding curve rate`}
-                tip={t`This rate determines the amount of overflow that each token can be redeemed for at any given time. On a lower bonding curve, redeeming a token increases the value of each remaining token, creating an incentive to hodl tokens longer than others. A bonding curve of 100% means all tokens will have equal value regardless of when they are redeemed.`}
+                tip={
+                  <Trans>
+                    This rate determines the amount of overflow that each token
+                    can be redeemed for at any given time. On a lower bonding
+                    curve, redeeming a token increases the value of each
+                    remaining token, creating an incentive to hodl tokens longer
+                    than others. A bonding curve of 100% means all tokens will
+                    have equal value regardless of when they are redeemed.
+                  </Trans>
+                }
               />
             }
           >
